Moved Cell, Microcell, Person, and IndexList into their own files to prevent circular dependencies due to their interaction with Dist.h
