+++ src/CMakeLists.txt
@@ -10,6 +10,7 @@ add_executable(CovidSim ${MAIN_SRC_FILES} ${MAIN_HDR_FILES})
 target_include_directories(CovidSim PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 
 add_subdirectory(Geometry)
+add_subdirectory(Models)
 
 if(USE_OPENMP)
   target_link_libraries(CovidSim PUBLIC OpenMP::OpenMP_CXX)
