+++ src/IndexList.h
@@ -1,0 +1,10 @@
+#pragma once
+
+/**
+ * @brief Used for computing spatial interactions more efficiently.
+ */
+struct IndexList
+{
+	int id;
+	float prob;
+};
