+++ src/Models/CMakeLists.txt
@@ -1,0 +1,6 @@
+target_sources(CovidSim
+    PRIVATE
+        ${CMAKE_CURRENT_LIST_DIR}/Cell.h
+        ${CMAKE_CURRENT_LIST_DIR}/Microcell.h
+        ${CMAKE_CURRENT_LIST_DIR}/Person.h
+)
